fix: AI tool appends phone/email/address by default, with replace option
- Update tool now defaults to appending new phone/email/address entries
  instead of silently replacing the primary entry (data loss bug)
- Add phoneAction/emailAction/addressAction params ("add" | "replace")
  so the AI model can express user intent correctly
- Return helpful guidance when search finds no results, suggesting
  shorter/partial queries to leverage Google's prefix matching
- Improve tool JSDoc with usage hints for the AI model
- Extract duplicated groupByLetter helper from ContactList/ContactGrid
- Fix ContactGrid to accept groups prop (was passed but silently ignored)
